requirements.txt
anyio==3.3.4
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
PyJWT==1.7.1
python-decouple==3.3
python-dotenv==0.19.1
PyYAML==6.0
regex==2021.10.23
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
